moved the user store into components/modules/user/stores
