+++ todo-app-standalone/src/components/common/Calender.vue
@@ -26,11 +26,11 @@
 
 <script>
 import { mapActions} from "pinia"
-import { userStore } from '../../stores/userStore'
+import { userStore } from '../modules/user/stores/userStore'
 
 import dayjs from 'dayjs'
-import PrimeCalender from "primevue/calendar"
-import Button from "primevue/button"
+import PrimeCalender from 'primevue/calendar'
+import Button from 'primevue/button'
 
 
 export default {
