+++ todo-app-standalone/src/components/modules/todo/components/TaskCompletedCard.vue
@@ -27,8 +27,8 @@
 
 <script>
 import { mapState, mapActions } from 'pinia'
-import { useTodoStore } from '@/stores/todoStore.js'
-import { userStore } from '../../../../stores/userStore'
+import { useTodoStore } from '@/stores/todoStore'
+import { userStore } from '../../user/stores/userStore'
 
 import CheckBox from '@/components/form/CheckBox.vue'
 import IconButton from '@/components/buttons/IconButton.vue'
