+++ todo-app-standalone/src/components/modules/user/components/RegisterPage.vue
@@ -39,7 +39,8 @@
 
 <script>
 import { mapActions } from 'pinia'
-import { userStore } from '../../../../stores/userStore'
+
+import { userStore } from '../../user/stores/userStore'
 
 import Password from 'primevue/password'
